porting router dummy app into tests/scenarios
