--- packages/router/ember-cli-build.js
@@ -1,23 +1,0 @@
-'use strict';
-
-const EmberAddon = require('ember-cli/lib/broccoli/ember-addon');
-
-module.exports = function (defaults) {
-  let app = new EmberAddon(defaults, {
-    // Add options here
-  });
-
-  if (process.env.CLASSIC) {
-    return app.toTree();
-  }
-
-  const Webpack = require('@embroider/webpack').Webpack;
-  return require('@embroider/compat').compatBuild(app, Webpack, {
-    staticAddonTestSupportTrees: true,
-    staticAddonTrees: true,
-    staticComponents: true,
-    staticHelpers: true,
-    staticModifiers: true,
-    splitAtRoutes: ['split-me'],
-  });
-};
