--- packages/router/testem.js
@@ -1,2 +1,0 @@
-const { testemConfig } = require('@embroider/test-support/testem-config');
-module.exports = testemConfig();
