--- packages/router/tests/dummy/app/app.js
@@ -1,12 +1,0 @@
-import Application from '@ember/application';
-import Resolver from 'ember-resolver';
-import loadInitializers from 'ember-load-initializers';
-import config from 'dummy/config/environment';
-
-export default class App extends Application {
-  modulePrefix = config.modulePrefix;
-  podModulePrefix = config.podModulePrefix;
-  Resolver = Resolver;
-}
-
-loadInitializers(App, config.modulePrefix);
