--- packages/router/tests/dummy/app/components/used-in-child.hbs
@@ -1,1 +1,0 @@
-<div data-test-used-in-child>This is the used-in-child component</div>
