--- packages/router/tests/dummy/app/controllers/split-me.js
@@ -1,3 +1,0 @@
-import Controller from '@ember/controller';
-
-export default class SplitMeController extends Controller {}
