--- packages/router/tests/dummy/app/controllers/split-me/child.js
@@ -1,3 +1,0 @@
-import Controller from '@ember/controller';
-
-export default class SplitMeChildController extends Controller {}
