--- packages/router/tests/dummy/app/routes/split-me/child.js
@@ -1,3 +1,0 @@
-import Route from '@ember/routing/route';
-
-export default class SplitMeChildRoute extends Route {}
