--- packages/router/tests/dummy/app/routes/split-me/index.js
@@ -1,3 +1,0 @@
-import Route from '@ember/routing/route';
-
-export default class SplitMeIndexRoute extends Route {}
