--- packages/router/tests/dummy/app/templates/application.hbs
@@ -1,9 +1,0 @@
-{{page-title 'Dummy'}}
-
-<h2 id='title'>Welcome to Ember</h2>
-
-<LinkTo @route='index'>Index</LinkTo>
-<LinkTo @route='split-me'>Split Index</LinkTo>
-<LinkTo @route='split-me.child'>Split Child</LinkTo>
-
-{{outlet}}
