--- packages/router/tests/dummy/app/templates/split-me.hbs
@@ -1,1 +1,0 @@
-{{outlet}}
