--- packages/router/tests/dummy/app/templates/split-me/child.hbs
@@ -1,1 +1,0 @@
-<UsedInChild />
