--- packages/router/tests/dummy/app/templates/split-me/index.hbs
@@ -1,1 +1,0 @@
-<div data-test-split-me-index>This is the split-me/index.</div>
