--- packages/router/tests/dummy/config/ember-cli-update.json
@@ -1,21 +1,0 @@
-{
-  "schemaVersion": "1.0.0",
-  "packages": [
-    {
-      "name": "ember-cli",
-      "version": "4.6.0",
-      "blueprints": [
-        {
-          "name": "addon",
-          "outputRepo": "https://github.com/ember-cli/ember-addon-output",
-          "codemodsSource": "ember-addon-codemods-manifest@1",
-          "isBaseBlueprint": true,
-          "options": [
-            "--yarn",
-            "--no-welcome"
-          ]
-        }
-      ]
-    }
-  ]
-}
