--- packages/router/tests/dummy/config/environment.js
@@ -1,49 +1,0 @@
-'use strict';
-
-module.exports = function (environment) {
-  let ENV = {
-    modulePrefix: 'dummy',
-    environment,
-    rootURL: '/',
-    locationType: 'history',
-    EmberENV: {
-      FEATURES: {
-        // Here you can enable experimental features on an ember canary build
-        // e.g. EMBER_NATIVE_DECORATOR_SUPPORT: true
-      },
-    },
-
-    APP: {
-      // Here you can pass flags/options to your application instance
-      // when it is created
-    },
-
-    isClassic: process.env.CLASSIC,
-  };
-
-  if (environment === 'development') {
-    // ENV.APP.LOG_RESOLVER = true;
-    // ENV.APP.LOG_ACTIVE_GENERATION = true;
-    // ENV.APP.LOG_TRANSITIONS = true;
-    // ENV.APP.LOG_TRANSITIONS_INTERNAL = true;
-    // ENV.APP.LOG_VIEW_LOOKUPS = true;
-  }
-
-  if (environment === 'test') {
-    // Testem prefers this...
-    ENV.locationType = 'none';
-
-    // keep test console output quieter
-    ENV.APP.LOG_ACTIVE_GENERATION = false;
-    ENV.APP.LOG_VIEW_LOOKUPS = false;
-
-    ENV.APP.rootElement = '#ember-testing';
-    ENV.APP.autoboot = false;
-  }
-
-  if (environment === 'production') {
-    // here you can enable a production-specific feature
-  }
-
-  return ENV;
-};
