--- packages/router/tests/dummy/config/optional-features.json
@@ -1,6 +1,0 @@
-{
-  "application-template-wrapper": false,
-  "default-async-observers": true,
-  "jquery-integration": false,
-  "template-only-glimmer-components": true
-}
