--- packages/router/tests/dummy/package.json
@@ -1,4 +1,0 @@
-{
-  "name": "dummy",
-  "version": "0.0.0"
-}
