--- packages/router/tests/test-helper.js
@@ -1,12 +1,0 @@
-import Application from 'dummy/app';
-import config from 'dummy/config/environment';
-import * as QUnit from 'qunit';
-import { setApplication } from '@ember/test-helpers';
-import { setup } from 'qunit-dom';
-import { start } from 'ember-qunit';
-
-setApplication(Application.create(config.APP));
-
-setup(QUnit.assert);
-
-start();
